Rename `op`→`wrapper_ops`
